requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
chromadb==0.4.22
openai>=1.10.0,<2.0.0
pypdf==3.17.3
python-dotenv==1.0.0